+++ src/snowflake/cli/plugins/spcs/common.py
@@ -75,7 +75,3 @@ def handle_object_already_exists(
         )
     else:
         raise error
-
-
-class NoPropertiesProvidedError(ValueError):
-    pass