+++ src/snowflake/cli/plugins/spcs/compute_pool/commands.py
@@ -10,7 +10,6 @@ from snowflake.cli.api.output.types import CommandResult, SingleQueryResult
 from snowflake.cli.api.project.util import is_valid_object_name
 from snowflake.cli.plugins.object.common import CommentOption
 from snowflake.cli.plugins.spcs.common import (
-    NoPropertiesProvidedError,
     validate_and_set_instances,
 )
 from snowflake.cli.plugins.spcs.compute_pool.manager import ComputePoolManager
@@ -123,74 +122,3 @@ def resume(name: str = ComputePoolNameArgument, **options) -> CommandResult:
     Resumes the compute pool from SUSPENDED state.
     """
     return SingleQueryResult(ComputePoolManager().resume(name))
-
-
-@app.command("set", requires_connection=True)
-def set_property(
-    name: str = ComputePoolNameArgument,
-    min_nodes: Optional[int] = MinNodesOption(default=None, show_default=False),
-    max_nodes: Optional[int] = MaxNodesOption(show_default=False),
-    auto_resume: Optional[bool] = AutoResumeOption(default=None, show_default=False),
-    auto_suspend_secs: Optional[int] = AutoSuspendSecsOption(
-        default=None, show_default=False
-    ),
-    comment: Optional[str] = CommentOption(
-        help="Comment for the compute pool.", show_default=False
-    ),
-    **options,
-) -> CommandResult:
-    """
-    Sets one or more properties or parameters for the compute pool.
-    """
-    try:
-        cursor = ComputePoolManager().set_property(
-            pool_name=name,
-            min_nodes=min_nodes,
-            max_nodes=max_nodes,
-            auto_resume=auto_resume,
-            auto_suspend_secs=auto_suspend_secs,
-            comment=comment,
-        )
-        return SingleQueryResult(cursor)
-    except NoPropertiesProvidedError:
-        raise ClickException(
-            f"No properties specified for compute pool '{name}'. Please provide at least one property to set."
-        )
-
-
-@app.command("unset", requires_connection=True)
-def unset_property(
-    name: str = ComputePoolNameArgument,
-    auto_resume: bool = AutoResumeOption(
-        default=False,
-        help="Reset the AUTO_RESUME property - The compute pool will automatically resume when a service or job is submitted to it.",
-        show_default=False,
-    ),
-    auto_suspend_secs: bool = AutoSuspendSecsOption(
-        default=False,
-        help="Reset the AUTO_SUSPEND_SECS property - Number of seconds of inactivity after which you want Snowflake to automatically suspend the compute pool.",
-        show_default=False,
-    ),
-    comment: bool = CommentOption(
-        default=False,
-        help="Reset the COMMENT property - Comment for the compute pool.",
-        callback=None,
-        show_default=False,
-    ),
-    **options,
-) -> CommandResult:
-    """
-    Resets one or more properties or parameters for the compute pool to their default value(s).
-    """
-    try:
-        cursor = ComputePoolManager().unset_property(
-            pool_name=name,
-            auto_resume=auto_resume,
-            auto_suspend_secs=auto_suspend_secs,
-            comment=comment,
-        )
-        return SingleQueryResult(cursor)
-    except NoPropertiesProvidedError:
-        raise ClickException(
-            f"No properties specified for compute pool '{name}'. Please provide at least one property to reset to its default value."
-        )