+++ RELEASE-NOTES.md
@@ -11,7 +11,7 @@
   * Added `image-registry login` command to fetch authentication token and log in to image registry in one command.
   * Added `image-repository url <repo_name>` command to get the URL for specified image repository.
   * Added `create` command for `image-repository`.
-  * Added `set (property)`, `unset (property)`, `suspend` and `resume` commands for `compute-pool`.
+  * Added `suspend` and `resume` commands for `compute-pool`.
   * Added `upgrade` and `list-endpoints` commands for `service`.
 
 ## Fixes and improvements